refactor files into folders
