--- src/durableerror.ts
@@ -1,7 +1,0 @@
-import * as types from "durable-functions";
-
-export class DurableError extends Error implements types.DurableError {
-    constructor(message: string | undefined) {
-        super(message);
-    }
-}
