+++ src/entities/entityid.ts
@@ -1,6 +1,6 @@
 // tslint:disable:member-access
 
-import { Utils } from "../utils";
+import { Utils } from "../util/Utils";
 import * as types from "durable-functions";
 
 export class EntityId implements types.EntityId {
