+++ src/http/httpcreationpayload.ts
@@ -1,0 +1,6 @@
+/** @hidden */
+export class HttpCreationPayload {
+    [key: string]: string;
+
+    constructor(public createNewInstancePostUri: string, public waitOnNewInstancePostUri: string) {}
+}
