--- src/httpcreationpayload.ts
@@ -1,6 +1,0 @@
-/** @hidden */
-export class HttpCreationPayload {
-    [key: string]: string;
-
-    constructor(public createNewInstancePostUri: string, public waitOnNewInstancePostUri: string) {}
-}
