+++ src/taskorchestrationexecutor.ts
@@ -11,11 +11,11 @@ import {
     TaskCompletedEvent,
     WaitForExternalEventAction,
 } from "./classes";
-import { OrchestrationFailureError } from "./orchestrationfailureerror";
+import { OrchestrationFailureError } from "./error/OrchestrationFailureError";
 import { OrchestratorState } from "./orchestratorstate";
 import { TaskBase, NoOpTask, DFTask, CompoundTask, TaskState } from "./task";
 import { ReplaySchema } from "./replaySchema";
-import { Utils } from "./utils";
+import { Utils } from "./util/Utils";
 import { DurableOrchestrationContext, OrchestrationContext } from "durable-functions";
 
 /**
